pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phishguard-vlm"
version = "0.1.0"
description = "Multimodal phishing detection with Vision-Language Models"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Lujain Khalil" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "torch>=2.0",
  "transformers>=4.36",
  "peft>=0.7",
  "accelerate>=0.25",
  "playwright>=1.40",
  "fastapi>=0.109",
  "uvicorn[standard]>=0.27",
  "requests",
  "aiohttp",
  "pandas",
  "pyarrow",
  "tqdm",
  "beautifulsoup4",
  "lxml",
  "pillow",
  "pyyaml",
  "wandb",
  "scikit-learn",
  "datasets",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "pytest-cov",
  "ruff",
  "mypy",
]
# Optional: screenshot OCR for cross-modal brand cues (also install system Tesseract)
ocr = ["pytesseract"]

[tool.setuptools.packages.find]
where = ["."]
include = ["configs*", "data_pipeline*", "models*", "knowledge_module*", "inference*", "evaluation*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
